Code fully implemented with hardware.
Implemented gameControls() function to describe position/movement of the analog joystick and appleEaten() function to activate a buzzer once the snake has eaten an apple. Serial.read() and Serial.write() functions used to send and receive bytes to and from the code in order for the python game "snake" to function with the hardware.